% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{checkDataConfig}
\alias{checkDataConfig}
\title{Check XT Configuration Columns}
\usage{
checkDataConfig(tbl, xt, if_error = stop, ...)
}
\arguments{
\item{tbl}{Data table (data frame) containing the actual data the shiny app
works with}

\item{xt}{A list representing the settings of a specific context mode}

\item{if_error}{Function that accepts a message
e.g. cat, message, warning, stop}

\item{...}{Arguments in ... will be passed to if_error()}
}
\value{
Logical. TRUE if XT config contains the nacessary info for tbl
}
\description{
Checks whether an XT config has
enough column info for a data table (data frame)
}
\seealso{
Other config functions: \code{\link{getConfigField}},
  \code{\link{getConfigMode}}
}
