% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{choices}
\alias{choices}
\title{Creates a "choices" vector for \code{shiny::selectizeInput()}}
\usage{
choices(var_col, .context, tbl_name = .context$tbl_name,
  keep_na = c("if_any", "omit_na", "insert_na"))
}
\arguments{
\item{var_col}{Name of the column to extract unique values}

\item{.context}{List. The shinyXT context.}

\item{tbl_name}{Name of the requested table.}

\item{keep_na}{How to handle missing values: \code{"if_any"} (default) 
keeps \code{NAs}, \code{"omit_na"} drops \code{NAs}, 
\code{"insert_na"} makes sure \code{NA} is the first value.}
}
\value{
A character vector
}
\description{
Creates a "choices" vector for \code{shiny::selectizeInput()}
}
