% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-misc.R
\name{createDT}
\alias{createDT}
\title{Create a Data Table.}
\usage{
createDT(.context)
}
\arguments{
\item{.context}{List. The shinyXT context.}
}
\value{
A DT DataTable
}
\description{
Creates a DT DataTable based on the settings specified in the context.
}
\examples{
onetable_df <- data.frame(
    a0_num = c(1, 2, 3, 4)
)

onetable_xt <- list(
    
    .default = list(
        col_name = NA_character_,
        class = NA_character_,
        ui_name = NA_character_,
        hover = NULL,
        widget = NULL,
        format = NULL,
        column_width = 3,   
        width = "100\%", 
        visible = TRUE,
        enabled = TRUE,
        validate = NULL
    ),
    
    .options = list(),
    a0_num = list(
        col_name = "a0_num",
        class = "numeric",
        ui_name = "Col A0",
        hover = "Numeric, no decimals",
        widget = "numericInput",
        format = "",
        validate = NULL
    )
)


context_default <- list(
    tbl_name = "onetable",
    tbl_lst = list(
        onetable = onetable_df
    ),
    filter_lst = NULL,
    mode = "dt",
    xt_lst = list(
        onetable = onetable_xt
    )
)

xt <- shinyXT::getConfigMode(context_default)

dt <- shinyXT::createDT(context_default)

}
\seealso{
\code{\link{formatDT}}
}
