% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{createForm}
\alias{createForm}
\title{Create New Record Form}
\usage{
createForm(.context)
}
\arguments{
\item{.context}{List. The shinyXT context.}
}
\description{
Creates the form that allows you to add a new record to your table,
 based on the setting in your context.
}
\seealso{
Other form functions: \code{\link{validateForm}}
}
\concept{form functions}
