% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-misc.R
\name{formatDT}
\alias{formatDT}
\title{Format a Data Table.}
\usage{
formatDT(dt, xt)
}
\arguments{
\item{dt}{A Data Table.}

\item{xt}{Config mode from context.}
}
\value{
the newly formatted DataTable
}
\description{
Formats the DT based on the options specified in the configuration.
}
\examples{
.XT <- list(
    valid_mode = c("dt", "edit", "add"),
    valid_class = c("numeric", "integer", "character", "Date", "POSIXct"),
    .default = list(),
    .options = list()
)
.context <- list(
    tbl_name = "onetable",
    mode = "edit",
    xt_lst = list(onetable = list(.XT = .XT))
)
xt <- getConfigMode(.context)

dt <- DT::datatable(
    data = data.frame(date = as.Date("2018/06/03")),
    options = list(format_Date = quote(format_utc),
                   format_POSIXct = quote(format_utc)),
    selection = 'single',
    rownames = FALSE,
    extensions = "Buttons",
    colnames = getConfigField(df, xt, 'ui_name'),
    filter = 'none',
    escape = -1
 ) \%>\%
    formatDT(xt)
}
\seealso{
\code{\link{createDT}}
}
