% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getConfigField}
\alias{getConfigField}
\title{Gets the Configuration Field}
\usage{
getConfigField(tbl, xt, field_name)
}
\arguments{
\item{tbl}{DT DataTable (data frame) containing the actual data the
shiny app works with}

\item{xt}{A list representing the settings of a specific context mode}

\item{field_name}{A character representing the config field}
}
\value{
A character representing the config field
}
\description{
Returns values from an xt field in the order of the \code{tbl} columns
}
\seealso{
Other config functions: \code{\link{checkDataConfig}},
  \code{\link{getConfigMode}}
}
\concept{config functions}
