% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{getConfigMode}
\alias{getConfigMode}
\title{Extracts Configuration Mode from Context}
\usage{
getConfigMode(.context, mode = .context$mode)
}
\arguments{
\item{.context}{List. The shinyXT context}

\item{mode}{The name of the config mode belonging to the context}
}
\value{
A list representing the settings of a specific context mode
}
\description{
Extracts from second level in xt_lst,
if it contains only valid modes.
If any level 2 (only!) field (e.g. validate) is a list of mode fields,
\code{getConfigMode()} will be used to extract only that mode.
}
\examples{
getConfigMode(.context = context, mode = .context$mode)

}
\seealso{
Other config functions: \code{\link{checkDataConfig}},
  \code{\link{getConfigField}}
}
\concept{config functions}
