% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{getEmptyRow}
\alias{getEmptyRow}
\title{Get an Empty Row}
\usage{
getEmptyRow(.context, tbl_name = .context$tbl_name)
}
\arguments{
\item{.context}{List. The shinyXT context.}

\item{tbl_name}{Name of the requested table.}
}
\description{
Returns a one row data frame of NA's.
}
