% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{getFullTbl}
\alias{getFullTbl}
\title{Get Entire Table}
\usage{
getFullTbl(.context, tbl_name = .context$tbl_name)
}
\arguments{
\item{.context}{List. The shinyXT context.}

\item{tbl_name}{Name of the requested table.}
}
\value{
The table from the contexts' table list.
}
\description{
Get Entire Table
}
\seealso{
\code{\link{getFilteredTbl}}
}
