% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{getNARow}
\alias{getNARow}
\title{Returns an one row data frame filled with NAs}
\usage{
getNARow(.context, tbl_name = .context$tbl_name)
}
\arguments{
\item{.context}{List. The shinyXT context.}

\item{tbl_name}{Name of the requested table.}
}
\value{
A data frame, preserving column classes
}
\description{
Returns an one row data frame filled with NAs
}
