% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{brXT}
\alias{brXT}
\alias{hrXT}
\title{Add a HTML tag}
\usage{
brXT()

hrXT()
}
\value{
A list with line break settings

A list with horizontal row settings
}
\description{
Adds a HTML Line Break

Adds a HTML horizontal row
}
