% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{isNotEmptyChr}
\alias{isNotEmptyChr}
\title{Check for NOT Empty Character}
\usage{
isNotEmptyChr(x)
}
\arguments{
\item{x}{Input to be checked.}
}
\value{
Logical: TRUE if NOT empty character and FALSE otherwise
}
\description{
Checks if input length is at least 1 and is not NA or empty string.
}
