% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-misc.R
\name{jsDT}
\alias{jsDT}
\title{Add JS scripts}
\usage{
jsDT(script_name = c("4col"))
}
\arguments{
\item{script_name}{the name of the script}
}
\value{
A string representing a JS script
}
\description{
Adds JS scripts to be used with Data Table calls.
}
