% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-misc.R
\name{progressInfo}
\alias{progressInfo}
\title{Displays the Progress Info}
\usage{
progressInfo(message, detail = "Begin", max = 10L, shiny_session = NULL,
  quiet = FALSE)
}
\arguments{
\item{message}{Character, the message to be displayed to the user.
\code{NULL} in order to hide the current message}

\item{detail}{Character, the detail to be displayed to the user. If the user
is in a Shiny session, detail message will be shown with a de-emphasized
appearance relative to message}

\item{max}{Numeric, value that represents the end of the progress bar}

\item{shiny_session}{The shiny session object, provided by shinyServer}

\item{quiet}{Logical, whether or not to display progress info}
}
\value{
A list with progress functions to call:
\item{name inc(detail)}{Increment progress counter} 
\item{name set(detail = NULL, value = NULL)}{Set progress counter 
  to \code{value}} 
\item{name close(detail = "Done")}{Terminate the progress bar. Best if called
  from \code{on.exit()}.}
}
\description{
Displays progress info taking into account the mode the user is in. If the
  \code{quiet} mode is activated, the function call doesn't display anything.
  If the user is in console mode, the progress information is printed on the
  console. If the function is called in a valid Shiny session,
  a default progress bar shows progress information.
}
