% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{shinyXTExample}
\alias{shinyXTExample}
\title{Create and run a shinyXT example}
\usage{
shinyXTExample(example_name = "onetable")
}
\arguments{
\item{example_name}{String that represents name of the example.}
}
\description{
Builds and runs the specified shinyXT example, based on the configuration
 files found in the same folder.
}
