% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{validateForm}
\alias{validateForm}
\title{Validate Form Based on a List of Input Views}
\usage{
validateForm(input_lst, .context)
}
\arguments{
\item{input_lst}{List of views to validate.}

\item{.context}{List. The shinyXT context.}
}
\value{
A list
}
\description{
Called when the user inserts values in the form and the input requires
  validation, based on the rules specified in the shinyXT context.
}
\examples{
observeEvent({
    input$my_input
}, {
    reactiveValuesToList(input) # get all the inputs on the page
    vld_lst <- validateForm(input_lst, .context)
})

}
\seealso{
Other form functions: \code{\link{createForm}}
}
\concept{form functions}
