% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-misc.R
\name{createDT}
\alias{createDT}
\title{Create a Data Table.}
\usage{
createDT(.context)
}
\value{
A DataTable
}
\description{
Creates a DT based on the settings specified in the context.
}
\seealso{
\code{\link{formatDT}}
}
