% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form.R
\name{createForm}
\alias{createForm}
\title{Create New Record Form}
\usage{
createForm(.context)
}
\arguments{
\item{.context}{List. The shinyXT context.}
}
\value{
A HTML tag object.
}
\description{
Creates the form that allows the addition of a new record to the table,
 based on the context settings.
}
\seealso{
Other form functions: \code{\link{validateForm}}
}
