% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-misc.R
\name{downloadButtonWithIcon}
\alias{downloadButtonWithIcon}
\title{Creates a Download Button with Icon}
\usage{
downloadButtonWithIcon(outputId, label = "Download", class = NULL,
  icon = shiny::icon("download"), ...)
}
\description{
Creates a download button or link which when clicked initiates a 
 browser download. It allows the user to customize the button by 
 adding an icon to it. It has a corresponding downloadHandler 
 in the server function.
 
 @param outputId Character, the name of the output slot to which the
  corresponding downloadHandler is assigned.
 @param label the label that appears on the button
 @param class additional CSS classes
 @param icon  the icon that appears on the button
 @param ... named attributes applied to the button
}
\seealso{
Other shiny misc functions: \code{\link{progressInfo}}
}
