% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny-misc.R
\name{downloadButtonWithIcon}
\alias{downloadButtonWithIcon}
\title{Creates a Download Button with Icon}
\usage{
downloadButtonWithIcon(outputId, label = "Download", class = NULL,
  icon = shiny::icon("download"), ...)
}
\arguments{
\item{outputId}{Character, the name of the output slot to which the
corresponding \code{downloadHandler} is assigned.}

\item{label}{the label that appears on the button}

\item{class}{additional CSS classes}

\item{icon}{the icon that appears on the button}

\item{...}{named attributes applied to the button}
}
\value{
A HTML tag object which can be also rendered as HTML
}
\description{
Has the same functionality as \code{shiny::downloadButton} but it also 
allows the user to customize the button by adding an icon to it.
It has a corresponding \code{downloadHandler}
in the server function.
}
\seealso{
\code{\link{shiny::downloadButton}}
}
