% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt-misc.R
\name{formatDT}
\alias{formatDT}
\title{Format a Data Table.}
\usage{
formatDT(dt, xt)
}
\arguments{
\item{dt}{Data table.}

\item{xt}{Config mode from context.}
}
\description{
Formats the DT based on the options specified in the configuration.
}
\examples{
xt <- getConfigMode(.context)
dt <- DT::datatable(
    df,
    options = options_lst,
    selection = .options$selection \%||\% 'single',
    rownames = FALSE,
    extensions = extensions,
    colnames = getConfigField(df, xt, 'ui_name'),
    filter = .options$filter \%||\% 'none',
    escape = -skip_escape_idx
 ) \%>\%
    formatDT(xt0)
    
}
\seealso{
\code{\link{validateDT}}
}
