% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xt-utils.R
\name{getFilteredTbl}
\alias{getFilteredTbl}
\title{Get Filtered Table}
\usage{
getFilteredTbl(.context, tbl_name = .context$tbl_name)
}
\arguments{
\item{.context}{List. The shinyXT context.}

\item{tbl_name}{Name of the requested table.}
}
\description{
Returns a filtered table from the contexts' table list and filter list.
}
\seealso{
@seealso \code{\link{getFilteredTbl}}
}
