% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{shinyXTExample}
\alias{shinyXTExample}
\title{Runs shinyXT examples included with the package}
\usage{
shinyXTExample(example_name = "onetable")
}
\arguments{
\item{example_name}{String that represents name of the example.}
}
\description{
Runs the specified shinyXT example within a shiny app, 
 based on the configuration files found in the same folder 
 (see folder \code{examples/}).
}
