% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{shinyXTExample}
\alias{shinyXTExample}
\title{based on https://deanattali.com/2015/04/21/r-package-shiny-app/}
\usage{
shinyXTExample(example_name = "onetable")
}
\description{
based on https://deanattali.com/2015/04/21/r-package-shiny-app/
}
